peer manager: Disable in tests
